24-2: progress
The try-all-directions idea was ok. I didn't get it to run fast enough
to actually brute force all numbers less than 500, but it evolved into a
manual gradient descent type thing, where I picked an objective function
and tried some different inputs by hand until I settled on a pretty good
one.

We're able to get the error (for some definition of error) down very low
(< 1%) but it's not enough to win. We need exact integer values.

The current best guesses are: (45, 306, 76) for the direction and
magnitude. And (233_740_160_691_462, 100_345_639_785_493,
230_473_854_139_001) for the starting point.

Inputting a few guesses on the website, we're now certain that the final
answer (sum of the components of `start`) is between 565T and 567T -- a
range of 3 trillion. Hey, it's something!

I think a promising next direction is to figure out what the heck a
gradient descent is, and see if/how it fits into what we're doing. I
have a good gut feeling about it, given the context (the other problems,
what we've tried so far, etc).